Add files via upload
********* Welcome To Guess Game *********

This is task1 project for UNIcompiler internship. #This game is written in python #Random,sympy modules are used in it. 1.Here the game is between player1 and player2. 2.Each player have 10 chances to guess a correct number and hints are provided at every wrong answer. 3.based on less guesses taken to correct guess is declared as winner. 4.game is tie,when both players taken same number of guesses.

********** Let's Play The Game **********